% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gene_enrichment.R
\name{Enrichment}
\alias{Enrichment}
\title{Gene enrichment analysis}
\usage{
Enrichment(res, id = "symbol", use = "GO", ont = "BP", fileName = NULL)
}
\arguments{
\item{res}{a character vector of gene IDs. Must be provided}

\item{id}{what type of ID is provided. "symbol" (default) or "entrezgene"}

\item{use}{which analysis to perform. "GO" (default) or "KEGG"}

\item{ont}{if "GO", which namespace to use. "MF", "BP" (default), of "CC". Not used in "KEGG" analysis}

\item{fileName}{save the results into a fileName. Default - none}
}
\value{
a data frame of top 20, or less, significant enrichments
}
\description{
A function to perform GO or KEGG enrichment analysis of a list of gene symbols or EntrezIDs
}
\note{
to visualize top 10 most significant results, use
if (nrow(res) > 10) n <-10 else n <- nrow(res)
grid.table(res[1:n, ], gp=gpar(fontsize=7))
}
\examples{
res <- Enrichment(gene_names_list, id="symbol", use="GO", , ont="MF")
res <- Enrichment(gene_names_list, id="entrezgene", use="KEGG", "results.txt")
}

