% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gene_enrichment.R
\name{gene_enrichment}
\alias{gene_enrichment}
\title{Performs GO or KEGG enrichment analysis of a list of gene symbols or Entrez IDs}
\usage{
gene_enrichment(selected, all.universe = NULL, id = "symbol", use = "GO",
  ont = "BP", fileName = NULL)
}
\arguments{
\item{selected}{a character vector of gene IDs. Either gene names or Entrez IDs are accepted. Required.}

\item{all.universe}{a character vector of all genes, to be used for estimating random enrichments.
Must be the same ID type as the 'selected' genes.}

\item{id}{what type of ID is provided. "symbol" (e.g., "BRCA1", default) or "entrezid" (e.g., "672").}

\item{use}{which analysis to perform. "GO" (default) or "KEGG".}

\item{ont}{if "GO", which ontology namespace to use. "MF", "BP" (default), of "CC". Not used in "KEGG" analysis}

\item{fileName}{save the results into a fileName. Default - none}
}
\value{
a data frame of significant enrichments, with lists of gene symbols per enriched funciton.
}
\description{
A function to perform GO or KEGG enrichment analysis of a list of gene symbols or EntrezIDs. A "background",
or the universe of all gene IDs can be provided instead of default all genes.
}
\note{
to visualize the top 10 most significant results, use
\code{if (nrow(res) > 10) { n <-10 } else { n <- nrow(res) }; kable(res[1:n, ])}
}
\examples{
\dontrun{
# Analysis of genes associated with asperger syndrome
res <- gene_enrichment(selected=c("100188800", "10849", "115727", "2272", "26059", "27185", "359778", "414", "4204", "431710", "431711", "449015", "4842", "4843", "50863", "6532", "79811", "80896", "831", "85358"), id="entrezid", use="GO", ont="MF")
res <- gene_enrichment(selected=c("DISC1", "ASPG4", "ASPG1", "ASPG2", "SLC6A4", "ASPG3", "FRAXE", "FRAXA", "FHIT", "NTM", "SLTM", "RASGRP4", "NOS2", "NOS1", "SHANK3", "DISC2", "TSNAX", "OXTR", "ARSD"), id="symbol", use="KEGG", fileName="results.txt")
}
}

