% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_filter.R
\name{gene_filter}
\alias{gene_filter}
\title{Filters expression matrix by two genefilter criteria}
\usage{
gene_filter(mtx, proportion = 0.33, average_expression = 10,
  min_difference = 2)
}
\arguments{
\item{mtx}{a matrix of genes (rows) vs. samples (columns), or an ExpressionSet. Required.}

\item{proportion}{a proportion of samples having gene expressed above 'average_expression'. Default - 0.33}

\item{average_expression}{expression level to be exceeded when calculating 'proportion'. Default - 10}

\item{min_difference}{expression difference between maximal and minimal intensity to be exceeded
for a gene to be used. Default - 2}
}
\value{
a filtered matrix
}
\description{
A wrapper of two genefilter functions to filter low exprssed and low variability genes (rows)
}
\note{
Example from "Statistics and Data Analysis for Microarrays Using R and Bioconductor" book by Sorin Drăghici, p/ 778

Filtering should be done with caution, 10.1186/1471-2105-11-450.
}
\section{TODO}{
 Implement three separate filtering functions: 1) Signal filtering, removing genes with expression
at or below defined level, at or more defined proportion; 2) Absolute value of change, max-min difference
should be larger than a defied threshold; 3) Variance filtering, variance should be larger than a defined threshold
}
\examples{

\dontrun{
set.seed(2)
mtx <- rbind(matrix(rnorm(200, mean = 9,  sd = 2), ncol = 20),
             matrix(rnorm(200, mean = 10, sd = 5), ncol = 20))
dim(mtx)
dim(gene_filter(mtx))
}

}

