% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pubmed_citations.R
\name{get_pubmed_citations}
\alias{get_pubmed_citations}
\title{Plots an XY plot of citation counts vs. year, colored by a category}
\usage{
get_pubmed_citations(pmids, log2_transform = TRUE,
  fileName = "get_pubmed_citations.pdf", legend_x = 0.8, legend_y = 0.8,
  fontSize_text = 4, fontSize_legend = 12)
}
\arguments{
\item{pmids}{a data frame with required three columns, "tools", "pmids", "category". 
The "tools" column contains the names of tools. The "pmids" column contains the corresponding PubMed IDs.
The "category" column contains which category a tool belong, used for coloring. Required}

\item{log2_transform}{whether to log2-transform citation counts. Useful to harmonize large range of counts. Default - TRUE}

\item{fileName}{name of the file to output 7.5x7.5in graph. PDF only. Default: "get_pubmed_citations.pdf"}

\item{legend_x, legend_y}{position of the legend. Default: 0.8, 0.8}

\item{fontSize_text, fontSize_legend}{font sizes for the tool names and legend text. Default: 4, 12}
}
\value{
plots a graph, saves it into the file, returns the "pmids" data frame annotated with years and citation counts
}
\description{
Idea - to visualize categorized software publications by year, citation counts.
For each tool (pmid), the function will pull up year and citation count, 
and plot them on one X (year) - Y (citations) plot, with tool names and
calored by categories.
}
\examples{
\dontrun{
# An example of the pmids data frame
pmids <- data.frame(tools = c("GREAT", "Enrichr", "Bedtools", "Bedtools", "Genomic HyperBrowser", "ChromHMM", "Segway"),
                    pmids = c("20436461", "23586463", "20110278", "25199790", "21182759", "20657582", "22426492"),
                    category = c("Gene-centric", "Gene-centric", "Genome-wide", "Genome-wide", "One-stop", "Machine-learning", "Machine-learning"))
# Or, make one in Excel, and copy-read from clipboard                   
pmids <- psych::read.clipboard.tab()

pmids_annotated <- get_pubmed_citations(pmids, legend_x = 0.7, legend_y = 1)
}
}
