% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pubmed_graph.R
\name{get_pubmed_graph}
\alias{get_pubmed_graph}
\title{Plot a barplot of counts a term appearing in PubMed}
\usage{
get_pubmed_graph(term, yearstart = 1970, yearend = 2016)
}
\arguments{
\item{term}{what to search. Required. Example: "reproducible research"}

\item{yearstart}{starting year to search the term. Required. Default: 1970}

\item{yearend}{last year to search the term. Required. Default: 2016}

\item{id}{what type of ID is provided. "symbol" (e.g., "BRCA1") or "entrezid" (e.g., "672", default and recommended).}
}
\value{
a ggplot object, and plot the barplot
}
\description{
A function to plot yearly counts of a term appearing in PubMed
}
\details{
Counts are not normalized by the total number of published papers
}
\examples{
\dontrun{
get_pubmed_graph("retraction", yearstart = 2000, yearend = 2016)
}
}

