% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_barplot.R
\name{gr_barplot}
\alias{gr_barplot}
\title{Internal. Plot a barplot of the enrichment results}
\usage{
gr_barplot(mtx, legend_location = "topright", bottom_margin = 5,
  x_axis_labels, pvalue_cutoff = 0.1)
}
\arguments{
\item{mtx}{a matrix to visualize as a barplot. Required}

\item{legend_location}{location to plot the legend. Common choices - "topright",
"bottomright". Default - "topright"}

\item{bottom_margin}{how much space to allocate for the X axis legend.
Default - 5}

\item{x_axis_labels}{a vector of labels to add to each bar under the X axis.}

\item{pvalue_cutoff}{where to draw dashed lines representing significance cutoff.}

\item{log10_transformed}{logical. Indicates the scale of the enrichment values.
Default - 0.1, 10 percent FDR.}
}
\value{
none
}
\description{
A function to plot a barplot of the enrichment results
}
\examples{
\dontrun{
gr_barplot(mtx = mtx.barplot.dn[, seq(1:length(use_columns)), drop=F], legend_location = "bottomright", bottom_margin = 8, x_axis_labels = names.args.dn, pvalue_cutoff = pvalue_cutoff)
}
}
