% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_cellspecific.R
\name{gr_cellspecific}
\alias{gr_cellspecific}
\title{Cell type-specific enrichment analysis}
\usage{
gr_cellspecific(mtx, cutoff.pval = 0.01, p2z = FALSE, fileName = NULL)
}
\arguments{
\item{mtx}{a matrix of the enrichment results (transformed p-values). The cell type-specific enrichment analysis is performed on each column (for each SNP set).}

\item{cutoff.pval}{a p-value cutoff for cell type enrichment significance. Default: 0.01}

\item{p2z}{logical. Indicates whether instead of -log10-transformation use
Z-scores. Default - FALSE.}

\item{fileName}{a path to a filename to save the results. Should be with "xlsx" extension. The results for each column are saved in separate worksheets. Default - NULL, do not save.}
}
\value{
A list of the SNP set-specific cell type enrichment results
}
\description{
A function to perform cell type-specific enrichment analysis on a matrix 
of the enrichment results (raw p-values). It compares the distributions 
of the overall and cell-specific -log10 p-values using Wilcoxon test.
Each column (SNP set-specific enrichment profile) is processed separately,
and the results are returned as a list contatining cell type enrichment
results for each set.
}
\examples{
\dontrun {
 gr_cellspecific(mtx)
}
}

