% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_clusters.R
\name{gr_clusters}
\alias{gr_clusters}
\title{Define clusters}
\usage{
gr_clusters(d, height = 10, minmembers = 3, fileName = NULL)
}
\arguments{
\item{d}{a dendrogram object. Required.}

\item{height}{a height threshold to cut the dendrogram. Should be defined empirically. Required, default - 10.}

\item{minmembers}{a minimum number of samples to be considered as a cluster. Required, default - 3.}

\item{fileName}{save the results into a fileName. Default - none}
}
\value{
a list with 'eset.labels' and 'eset.groups' slots defining clustered labels/groups 
from the dendrogram cutting save the clustering order in a file
}
\description{
A function to define clusters by cutting a dendrogram by specified height.
}
\examples{
\dontrun{
mtx.clust <- gr_clusters(d=h$Colv, height=1.5, minmembers=3, fileName="clustering.txt")
}
}

