% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_gene_extract.R
\name{gr_gene_extract}
\alias{gr_gene_extract}
\title{Extract gene regions for the list of genes}
\usage{
gr_gene_extract(selected, id = "entrezid")
}
\arguments{
\item{selected}{a character vector of gene Entrez IDs, or names. Required}

\item{id}{what type of ID is provided. "symbol" (e.g., "BRCA1") or "entrezid" (e.g., "672", default and recommended).}
}
\value{
a list with two components. 'genes' contains genomic coordinates
of the selected genes; 'notfound' contains gene names not found in annotables (for diagnostics)
}
\description{
A function to extract genomic coordinates of genes in BED format.
}
\details{
Only canonical chromosome names are considered. No patches, random or other contigs.
}
\examples{
\dontrun{
selected <- c("TMEM59L", "SNTG1", "RPL41", "ADAMTS19")
genes_bed <- gr_gene_extract(selected, id = "symbol")$genes
write.table(genes_bed, "genes.bed", sep="\\t", quote=F, row.names=F, col.names=F)
# Extract genomic coordinates of all genes. Note only canonical chromosomes are retained.
genes_all <- gr_gene_extract(selected = unique(annotables::grch37$symbol), id = "symbol")
}
}
