% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_getV1OddsRatioPvalMatrix.R
\name{gr_getV1OddsRatioPvalMatrix}
\alias{gr_getV1OddsRatioPvalMatrix}
\title{A function to convert V1 to V2 GenomeRunner results}
\usage{
gr_getV1OddsRatioPvalMatrix(infile)
}
\arguments{
\item{infile}{path to the log file. Required}
}
\value{
generates two files, "matrix_PVAL.txt" and "matrix_OR.txt" in the same
filder as the log file
}
\description{
Converts a log file of the results obtained with GenomeRunner V1 (standalone)
to the "matrix_PVAL/OR.txt" files compatible with GenomeRunner V2 (web)
}
\note{
Courtesy to \link[https://github.com/AmyOlex]{Amy Olex}
}
\examples{
\dontrun{
gr_getV1OddsRatioPvalMatrix("TFBS100_LOG.gr")
}
}
