% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_load_data.R
\name{gr_load_data}
\alias{gr_load_data}
\title{Load enrichment analysis data}
\usage{
gr_load_data(dname, subset = "none", p2z = FALSE)
}
\arguments{
\item{dname}{a string, or a character vector containing one or multiple paths 
to the matrix file(s). Multiple matrixes will be 'rbind'. If a matrix name 
has 'PVAL' in its name, the data will be -log10-transformed. If a matrix name 
has 'OR' in its name, the data will be log2-transformed. If no such keywords 
are found, the data is not transformed. The non-informative enrichments 
(rows with all non-significant enrichments) are removed}

\item{subset}{a string used to select rows containing it. If a vector, "OR"
operation is used in subsetting. Default - none. Examples - c("Histone", "Tfbs")}

\item{p2z}{logical. Indicates whether instead of -log10-transformation use
Z-scores. Default - FALSE.}
}
\value{
a matrix of transformed data
}
\description{
A function to load enrichment analysis matrix(es) and remove non-informative 
enrichment results.
}
\examples{
\dontrun{
mtx <- gr_load_data("data/ENCODE/matrix_OR.txt")
mtx <- gr_load_data(c("data/ENCODE_Tfbs/matrix_PVAL.txt", 
"data/ENCODE_Histone/matrix_PVAL.txt"), subset=c("Tfbs", "Histone"))
}
}

