% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gr_maxmin.R
\name{gr_maxmin}
\alias{gr_maxmin}
\title{Extract max/min correlations for pairs of objects (columns) from a matrix}
\usage{
gr_maxmin(mtx, cortype = "spearman", fileName = NULL)
}
\arguments{
\item{mtx}{a n x m matrix used to create an m x m square matrix of correlation
coefficients (column-wise correlations). Required}

\item{cortype}{method to calculate correlation coefficients. Default - "spearman".
Other oprions - "pearson".}

\item{fileName}{name of a tab-separated text file to save the results.
Default - none, do not save the results}
}
\value{
an m x 7 matrix with the 7 columns:
\preformatted{
- obj - target object
- obj.max - object maximally correlated with the target object
- obj.min - object minimally correlated with the target object
- corr.coeff.max - max correlation coefficient
- corr.coeff.min - min correlation coefficient
- p.value.max - p-value for max correlation coefficient
- p.value.min - p-value for min correlation coefficient
}
}
\description{
A function to extract max/min correlation of each object (column) with other
objects. Each column in an n x m matrix is correlated with other columns, and
the names of the columns maximally/minimally correlated with the target column
are assembled
}
\examples{
\dontrun{
mtx.maxmin <- gr_maxmin(mtx, cortype="pearson", fileName = "maxmin.txt")
}
}

