% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_plot.R
\name{gr_plot}
\alias{gr_plot}
\title{Visualize enrichment results}
\usage{
gr_plot(file_name, use_columns = 1, subset_by_GF = "none",
  subset_by_cell = "none", log10_transformed = FALSE,
  adjust_pval_by = "fdr", pvalue_cutoff = 0.1, numtofilt = 1,
  plot_type = "bar")
}
\arguments{
\item{file_name}{path to the file with the enrichment results. Required}

\item{use_columns}{which column(s) to use. Default - 1, use the first column}

\item{subset_by_GF}{a string expression that is used to subset rows (GFs). 
Multiple strings OR allowed, e.g. c("H3K4me1", "H3K27ac"). Non-case sensitive.
Default - "none", use all rows}

\item{subset_by_cell}{a string expression used to subset rows (GFs) by
cell type-specific GFs. Should be ENCODE or Roadmap IDs. Multiple strings
are OR allowed, e.g., c("Gm12878", "K562"), c("E033", "E039). Non-case
sensitive. Devault - "none", use all rows}

\item{log10_transformed}{logical. Indicates the scale of the enrichment values.
Used to properly convert the values for visualization. If "OR", values are
treated as odds ratios. Default - FALSE, values are regular p-values.}

\item{adjust_pval_by}{method to adjust p-values. Use the same methods as
accepted by the 'p.adjust' function. Common choices - "none", "fdr". Default -
"fdr"}

\item{pvalue_cutoff}{threshold below which results are not considered. 
Default - 0.1, or 10 percent FDR.}

\item{numtofilt}{needs refactoring}

\item{plot_type}{what to plot. "heat" is relevant only for one-column enrichment
matrix, "barup", "bardn", "bar" is used for plotting barplots of enrichment,
depletion, or both directions results. Default - "bar"}
}
\value{
a list with enrichment values used for plotting
}
\description{
A master function that visualizes enrichment results as heatmaps or barplots
}
\note{
Adapted from 'utils2.R'. Currently "bar" plotting has been tested.
}
\examples{
\dontrun{
# Barplot
gr_plot(file_name = "matrix_PVAL.txt", use_columns = 1, subset_by_GF = "none", subset_by_cell = "none", log10_transformed = FALSE, adjust_pval_by = "fdr", pvalue_cutoff = 0.1, numtofilt = 1, plot_type = "bar")
}
}

