% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gr_transform.R
\name{gr_transform}
\alias{gr_transform}
\title{Conversion of a matrix of p-values to -log10 scale}
\usage{
gr_transform(mtx, p2z = FALSE)
}
\arguments{
\item{mtx}{a matrix of p-values, with "-" sign indicating depleted enrichments. Required.}

\item{p2z}{logical. Indicates whether instead of -log10-transformation use
Z-scores. Default - FALSE.}
}
\value{
a matrix of -log10-transformed p-values, with "-" sign preserved
}
\description{
A function to convert a matrix of p-values (with "-" indicating depletion)
into a -log10-transformed matrix with sign preserved
}
\examples{
\dontrun{
mtx.transformed <- gr_transform(mtx)
}
}

