% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gr_trimnames.R
\name{gr_trimnames}
\alias{gr_trimnames}
\title{Trimming long column/row names}
\usage{
gr_trimnames(labs, num.char = 20)
}
\arguments{
\item{labs}{a character vector of names to trim. Required}

\item{num.char}{a number specifying length of a name after which the name
 is trimmed and "..." added. Default - 20 characters.}
}
\value{
a character vactor of trimmed names.
}
\description{
A function to trim long column/row names to a defined length.
}
\examples{
\dontrun{
colnames(mtx) <- gr_trimnames(colnames(mtx))
}
}

