% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_func.R
\name{pca_func}
\alias{pca_func}
\title{A function for PCA plotting with highlighting groups using ellipses}
\usage{
pca_func(data, groups, title, print_ellipse = TRUE)
}
\arguments{
\item{data}{a gene-by-sample numeric matrix. Required}

\item{groups}{a character vector of sample group assignment. Required}

\item{title}{a title for the graph. Required}

\item{print_ellipse}{highlight groups by ellipses, Default: TRUE}
}
\value{
ggplot2 object
}
\description{
Beautiful PCA plot
}
\note{
Source: \url{https://shiring.github.io/machine_learning/2017/01/15/rfe_ga_post}
}
\examples{
\dontrun{
library(pcaGoPromoter)
library(ellipse)
library(gridExtra)
library(grid)
library(ggplot2)
p1 <- pca_func(data = t(mtcars), groups = as.character(mtcars$am), print_ellipse = TRUE, title = "mtcars samples colored by transmission type")
p2 <- ggplot(data.frame(prcomp(mtcars, scale. = TRUE, center = TRUE)$x, Transmission = factor(mtcars$am)), aes(x = PC1, y = PC2, color = Transmission)) + geom_point(shape = 1)
grid.arrange(p1, p2, ncol = 2)
}
}
