% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_enrichr.R
\name{save_enrichr}
\alias{save_enrichr}
\title{A function to run enrichment analysis using \link[http://amp.pharm.mssm.edu/Enrichr/]{EnrichR} API
and, optionally, save them in an Excel file}
\usage{
save_enrichr(up.genes = NULL, dn.genes = NULL, databases = "KEGG_2016",
  fdr.cutoff = 1, fileName = NULL, wb = NULL, sheetName = NULL)
}
\arguments{
\item{up.genes}{a vactor of gene names defined as upregulated. Required}

\item{dn.genes}{a vactor of gene names defined as downregulated. Optional}

\item{databases}{database name in EnrichR. See help for \link[enrichR]{enrichGeneList} function. Default: 'KEGG_2016'}

\item{fdr.cutoff}{an FDR cutoff for enrichment results to be considered significant, Default: 1 (output all)}

\item{fileName}{a name of an Excel file to save the results, Default: NULL (do not save)}

\item{wb}{a workbook created beforehand. See \link{save_res} function. Default: NULL (do not save)}

\item{sheetName}{a name of a worksheet to save the results. Default: NULL (use `databases` as a worksheet name). If provided, takes priority over `datavases`}
}
\value{
a data frame with the enrichment results. And, if fileName etc. are provided, saves them
}
\description{
Given a vector of upregulated gene names (symbols), or two vectors of up- and downregulated genes,
run enrichment analysis using \link[https://github.com/compbiomed/enrichR]{enrichR} package. Gene lists
are analyzed separately, so for two gene lists the results are simply concatenated into one data frame,
with "UP" and "DN" indicator of enrichment results corresponding to a given list.
}
\examples{
\dontrun{
library(openxlsx)
fileName <- "results/DEGs.xlsx"
unlink(fileName) # Delete previous file
wb <- openxlsx::createWorkbook(fileName) # openxlsx::loadWorkbook(fileName) # Or, load existing
# Enrichment analysis of up/downregulated genes. Saves results directly into Excel file, and returns as a data frame
res <- save_enrichr(up.genes = up.genes, dn.genes = dn.genes, databases = "KEGG_2016", fdr.cutoff = fdr.cutoff, fileName = fileName, wb = wb)
DT::datatable(res) # Visualize
}

}
