% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_res.R
\name{save_res}
\alias{save_res}
\title{A function to save a data frame or a matrix into an Excel file. Uses \code{\link[openxlsx]{openxlsx}}}
\usage{
save_res(res, fileName = "fileName.xlsx", wb = wb, sheetName = "KEGG")
}
\arguments{
\item{res}{a data frame or a matrix to be saved}

\item{fileName}{a file name with `.xlsx` extension to save the results, Default: 'fileName.xlsx'}

\item{wb}{a workbook object, created initially, Default: wb}

\item{sheetName}{a name of a worksheet to save the results, Default: 'KEGG'}
}
\value{
nothing, just saves the results
}
\description{
Given a `fileName`, create new (\code{wb <- openxlsx::createWorkbook(fileName)})
or load existing (\code{wb <- openxlsx::loadWorkbook(fileName)}) Excel file. Then, use `save_res`
to save `res` objects into this file into a specified worksheet. One file can hold multiple worksheets.
No need to explicitly close the file.
}
\examples{
\dontrun{
library(openxlsx)
fileName <- "results/DEGs.xlsx"
unlink(fileName) # Delete previous file
wb <- openxlsx::createWorkbook(fileName) # openxlsx::loadWorkbook(fileName) # Or, load existing
res <- enrichR::enrichFullGeneList(up.genes, dn.genes, databases = "KEGG_2016", fdr.cutoff = 1) # Enrichment analysis of up/downregulated genes
save_res(res, fileName = fileName, wb = wb, sheetName = "KEGG")
}
}
