% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_corr.R
\name{TCGA_corr}
\alias{TCGA_corr}
\title{Retrieves correlation statistics between a gene of interest and all other genes.}
\usage{
TCGA_corr(cancer = "BRCA", gene = "BRCA1")
}
\arguments{
\item{cancer}{- cancer type abbreviation. Required. Example: "BRCA". One of
c("ACC", "BLCA", "BRCA", "CESC", "CHOL", "COAD", "COADREAD", "DLBC", "ESCA", "FPPP", 
"GBM", "GBMLGG", "HNSC", "KICH", "KIPAN", "KIRC", "KIRP", "LAML", "LGG", "LIHC", 
"LUAD", "LUSC", "MESO", "OV", "PAAD", "PCPG", "PRAD", "READ", "SARC", "SKCM", "STAD", 
"TGCT", "THCA", "THYM", "UCEC", "UCS", "UVM"). Look up
abbreviations at \url{http://www.liuzlab.org/TCGA2STAT/CancerDataChecklist.pdf}
and \url{https://tcga-data.nci.nih.gov/datareports/codeTablesReport.htm}}

\item{gene}{gene of interest. Required. Example: "BRCA1"}
}
\value{
a sorted by correlation coefficient data frame. Columns are 
"hgnc_symbol", "cor", "pval", "description"
}
\description{
A function to get Pearson correlation coefficients between RSEM gene expression of
a gene of interest and all other genes in a specific cancer type.
}
\examples{

\dontrun{
# Get correlation matrixes for the gene of interest in all cancers
cancers <- c("ACC", "BLCA", "BRCA", "CESC", "CHOL", "COAD", "COADREAD", "DLBC", "ESCA", "FPPP", "GBM", "GBMLGG", "HNSC", "KICH", "KIPAN", "KIRC", "KIRP", "LAML", "LGG", "LIHC", "LUAD", "LUSC", "MESO", "OV", "PAAD", "PCPG", "PRAD", "READ", "SARC", "SKCM", "STAD", "TGCT", "THCA", "THYM", "UCEC", "UCS", "UVM")
gene <- "FLI1"

for (i in cancers) {
  print(i)
  gene_corr <- TCGA_corr(cancer = i, gene = gene)
  write.table(gene_corr, paste0("TCGA_corr_", gene, "_", i, ".txt"), sep = "\\t", quote = F, row.names = F)
}
}
}

