% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unembed.R
\name{unembed}
\alias{unembed}
\title{Extracts multiple variables embedded in rows}
\usage{
unembed(data, col, sep, ...)
}
\arguments{
\item{data}{a data.frame object, with a column containing values with multiple
variables.}

\item{col}{column name containing values with multiple variables.}

\item{sep}{regular expression to split column by}
}
\value{
a data frame with columns containing unique variables obtained by
unembedding.
}
\description{
Extracts multiple values embedded in rows. E.g. one row with ABC11 /// BCD22
variable will be split into two separate entries, creating two ABC11 and BCD22
rows with other values equal to the original row.
}
\note{
Courtesy to \link[https://github.com/aaronwolen]{Aaron Wolen}
}
\examples{
\dontrun{
df <- data.frame(key = c("a", "a;b", "a;b;c"), val = 1:3)
unembed(df, "key", ";")
}
}

