% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{Venn3}
\alias{Venn3}
\title{Make three-circle Venn diagram}
\usage{
Venn3(set1, set2, set3, names = c("set 1", "set 2", "set 3"),
  title = "Venn diagram")
}
\arguments{
\item{set1}{a character vector of values in the 1st set. Required}

\item{set2}{a character vector of values in the 2nd set. Required}

\item{set3}{a character vector of values in the 3rd set. Required}

\item{names}{a character vector of length two containing names of the sets. Default: c("set 1", "set 2", "set 3")}

\item{title}{title of the Venn diagram. Default: "Venn diagram".}
}
\value{
a data frame with four columns used to get overlap counts
}
\description{
Make three-circle Venn diagram
}
\examples{
\dontrun{
#define some sets
e <- c(1, 2, 3, 4, 5)
f <- c(3, 4, 5, 6, 7)
g <- c(1, 7, 8)
Venn3(e, f, g, c("e", "f", "g"), "E, F and G")
}
}

