% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{Venn4}
\alias{Venn4}
\title{Make four-circle Venn diagram}
\usage{
Venn4(set1, set2, set3, set4, names = c("set 1", "set 2", "set 3", "set 4"),
  title = "Venn diagram")
}
\arguments{
\item{set1234}{a character vector of values in the nth set. Required}
}
\value{
Nothing, just plot the diagram
}
\description{
Make four-circle Venn diagram
}
\note{
Source \link[http://bioinfo-mite.crb.wsu.edu/Rcode/Venn2.R]{dead link}
}
\examples{
\dontrun{
# define some sets
set1 = letters[1:6];
set2 = letters[6:10];
set3 = letters[6:15];
set4 = letters[6:21];
Venn4(set1, set2, set3, set4);
}
}
