% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{Venn5}
\alias{Venn5}
\title{Make five-circle Venn diagram}
\usage{
Venn5(set1, set2, set3, set4, set5)
}
\arguments{
\item{set12345}{a character vector of values in the nth set. Required}
}
\value{
Nothing, just plot the diagram
}
\description{
Make five-circle Venn diagram
}
\note{
Source \link[http://bioinfo-mite.crb.wsu.edu/Rcode/Venn2.R]{dead link}
}
\examples{
\dontrun{
# define some sets
set1 = letters[1:6];
set2 = letters[6:10];
set3 = letters[6:15];
set4 = letters[6:21];
set5 = letters[6:26];
Venn5(set1, set2, set3, set4, set5);
}
}
