% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_enrichment.R
\name{gene_enrichment}
\alias{gene_enrichment}
\title{Performs GO or KEGG enrichment analysis of a list of gene symbols or Entrez IDs}
\usage{
gene_enrichment(selected, all.universe = NULL, id = "symbol", use = "GO",
  ont = "BP", pval = 0.05, p.adj = 0.1, fileName = NULL)
}
\arguments{
\item{selected}{a character vector of gene IDs. Either gene names or Entrez IDs are accepted. Required.}

\item{all.universe}{a character vector of all genes, to be used for estimating random enrichments.
Must be the same ID type as the 'selected' genes.}

\item{id}{what type of ID is provided. "symbol" (e.g., "BRCA1", default) or "entrezid" (e.g., "672").}

\item{use}{which analysis to perform. "GO" (default) or "KEGG".}

\item{ont}{if "GO", which ontology namespace to use. "MF", "BP" (default), of "CC". Not used in "KEGG" analysis}

\item{pval}{non-corrected p-value cutoff to perform the enrichment analysis. Default - 0.05.}

\item{p.adj}{FDR-corrected p-value cutoff to report the enrichment analysis results. Default - 0.1.}

\item{fileName}{save the results into a fileName. Default - none}
}
\value{
a data frame of significant enrichments, with lists of gene symbols per enriched funciton.
}
\description{
A function to perform GO or KEGG enrichment analysis of a list of gene symbols or EntrezIDs. A "background",
or the universe of all gene IDs can be provided instead of default all genes.
}
\note{
to visualize the top 10 most significant results, use
\code{if (nrow(res) > 10) { n <-10 } else { n <- nrow(res) }; kable(res[1:n, ])}
}
\examples{

\dontrun{
# Analysis of genes associated with asperger syndrome
res <- gene_enrichment(selected=c("100188800", "10849", "115727", "2272", "26059", "27185", "359778", "414", "4204", "431710", "431711", "449015", "4842", "4843", "50863", "6532", "79811", "80896", "831", "85358"), id="entrezid", use="GO", ont="MF")
res <- gene_enrichment(selected=c("DISC1", "ASPG4", "ASPG1", "ASPG2", "SLC6A4", "ASPG3", "FRAXE", "FRAXA", "FHIT", "NTM", "SLTM", "RASGRP4", "NOS2", "NOS1", "SHANK3", "DISC2", "TSNAX", "OXTR", "ARSD"), id="symbol", use="KEGG", fileName="results.txt")

# A wrapper function to perform all functional enrichment analyses. The 'fileName' should have 'xlsx' extension
all.enrichment.analyses <- function(genes, fileName, id="symbol") {
  # Gene ontology, molecular function
  res <- gene_enrichment(selected = genes, id=id, use="GO", ont="MF")
  write.xlsx(res, fileName, sheetName = "GO-MF", row.names=FALSE, append=TRUE)
  # Gene ontology, biological process 
  res <- gene_enrichment(selected = genes, id=id, use="GO", ont="BP")
  write.xlsx(res, fileName, sheetName = "GO-BP", row.names=FALSE, append=TRUE)
  # Gene ontology, cellular component
  res <- gene_enrichment(selected = genes, id=id, use="GO", ont="CC")
  write.xlsx(res, fileName, sheetName = "GO-CC", row.names=FALSE, append=TRUE)
  # KEGG canonical pathways
  res <- gene_enrichment(selected = genes, id=id, use="KEGG")
  write.xlsx(res, fileName, sheetName = "KEGG", row.names=FALSE, append=TRUE)
}

# To perform Reactome and Disease Ontology enrichment analyses, use the corresponding packages
library("ReactomePA")
library("DOSE")
res <- enrichPathway(gene = entrez.genes, universe = all.entrez, organism = "human", pAdjustMethod = "BH", pvalueCutoff = 0.1, qvalueCutoff = 0.2, readable = TRUE)
write.xlsx(summary(res), fileName, sheetName = "Reactome",row.names=FALSE,  append = TRUE)
res <- enrichDO(gene = entrez.genes, universe = all.entrez, ont = "DO", pAdjustMethod = "none", pvalueCutoff = 0.1, qvalueCutoff = 0.2, readable = TRUE)
write.xlsx(summary(res), fileName, sheetName = "DiseaseOntology",row.names=FALSE,  append = TRUE)
}

}

