% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_enrichment.R
\name{gene_enrichment}
\alias{gene_enrichment}
\title{Performs GO or KEGG enrichment analysis of a list of gene symbols or Entrez IDs}
\usage{
gene_enrichment(selected, all.universe = NULL, id = "symbol",
  organism = "Hs", use = "GO", ont = "BP", pval = 0.05, p.adj = 0.1,
  fileName = NULL)
}
\arguments{
\item{selected}{a character vector of gene IDs. Either gene names or Entrez IDs are accepted. Required.}

\item{all.universe}{a character vector of all genes, to be used for estimating random enrichments.
Must be the same ID type as the 'selected' genes.}

\item{id}{what type of ID is provided. "symbol" (e.g., "BRCA1", default) or "entrezid" (e.g., "672").}

\item{organism}{organism ID. "Hs" (homo sapiens, default), or "Mm" (mus musculus)}

\item{use}{which analysis to perform. "GO" (default), "KEGG", or "msigdf" for use with https://github.com/stephenturner/msigdf.}

\item{ont}{if "GO", which ontology namespace to use: "MF", "BP" (default), of "CC". Not used in "KEGG" analysis. 
If "msigdf", should be one of "c1": positional gene sets, "c2": curated gene sets, 
"c3": motif gene sets, "c4": computational gene sets, "c5": GO gene sets, 
"c6": oncogenic signatures, "c7": immunologic signatures, "hallmark": hallmark gene sets.
See http://software.broadinstitute.org/gsea/msigdb/ for more details.}

\item{pval}{non-corrected p-value cutoff to perform the enrichment analysis. Default - 0.05.}

\item{p.adj}{FDR-corrected p-value cutoff to report the enrichment analysis results. Default - 0.1.}

\item{fileName}{save the results into a fileName. Default - none}
}
\value{
a data frame of significant enrichments, with lists of gene symbols per enriched funciton.
}
\description{
A function to perform GO or KEGG enrichment analysis of a list of gene symbols or EntrezIDs. A "background",
or the universe of all gene IDs can be provided instead of default all genes.
}
\note{
to visualize the top 10 most significant results, use
\code{if (nrow(res) > 10) { n <-10 } else { n <- nrow(res) }; kable(res[1:n, ])}
}
\examples{

\dontrun{
# Analysis of genes associated with asperger syndrome
genes <- c("100188800", "10849", "115727", "2272", "26059", "27185", "359778", "414", "4204", "431710", "431711", "449015", "4842", "4843", "50863", "6532", "79811", "80896", "831", "85358")
res <- gene_enrichment(selected = genes, id = "entrezid", organism = "Hs", use = "GO", ont = "MF")
genes <- c("DISC1", "ASPG4", "ASPG1", "ASPG2", "SLC6A4", "ASPG3", "FRAXE", "FRAXA", "FHIT", "NTM", "SLTM", "RASGRP4", "NOS2", "NOS1", "SHANK3", "DISC2", "TSNAX", "OXTR", "ARSD")
res <- gene_enrichment(selected = genes, id = "symbol", organism = "Hs", use = "KEGG", fileName = "results.txt")
genes <- c("4794", "25816", "9516", "19", "3601", "9308", "10950", "9120", "50486", "6446", "6890", "4609", "1435", "4170", "3383", "2354", "3164", "8870", "22822", "1846", "80176", "9314", "3606", "604", "7280", "8613", "3575", "10611", "6347", "80149", "150094", "595", "3659", "57007", "7128", "329", "1052", "2152", "1844", "1958", "3553", "10769", "2526", "4084", "597", "1960", "9334", "1847", "23135", "56937", "3326", "619279", "105371473", "9245", "100271380", "54541", "9666", "102724919", "105377252", "347252", "100873354", "6987", "386665", "84912", "105372408", "643406", "105372132", "105372270", "105371183", "574037", "102725016", "100500920", "101928274", "46", "693213", "55751", "360001", "100421721", "283693", "8411", "65095", "100873516", "408187", "105370802", "7544", "497048", "106481539", "105377107", "105376710", "27092", "9874", "100616418", "254783", "100653365", "152485", "28951", "105370436", "6902", "651258", "100129424")
res <- gene_enrichment(selected = genes, id = "entrezid", use = "msigdf", ont = "hallmark")

# A wrapper function to perform all functional enrichment analyses.
# Helper function to save non-empty results
save_res <- function(res, fileName = fileName, wb = wb, sheetName = "KEGG") {
  if (nrow(res) > 0) {
    openxlsx::addWorksheet(wb = wb, sheetName = sheetName)
    openxlsx::writeData(wb, res, sheet = sheetName)
    openxlsx::saveWorkbook(wb, fileName, overwrite = TRUE)
  }
}
# Create (or, load)  Excel file
fileName <- "fileName.xlsx"
wb <- openxlsx::createWorkbook(fileName) # openxlsx::loadWorkbook(fileName)
# Gene ontology, molecular function
res <- gene_enrichment(selected = genes, id="symbol", organism = "Hs", use="GO", ont="MF")
save_res(res, fileName, wb = wb, sheetName = "GOMF")
# Gene ontology, biological process 
res <- gene_enrichment(selected = genes, id="symbol", organism = "Hs", use="GO", ont="BP")
save_res(res, fileName, wb = wb, sheetName = "GOBP")
# Gene ontology, cellular component
res <- gene_enrichment(selected = genes, id="symbol", organism = "Hs", use="GO", ont="CC")
save_res(res, fileName, wb = wb, sheetName = "GOCC")
# KEGG canonical pathways
res <- gene_enrichment(selected = genes, id="symbol", organism = "Hs", use="KEGG")
save_res(res, fileName, wb = wb, sheetName = "KEGG")

# Legend for GO/KEGG functional enrichment results: "ID" - unique identifier of functional category; 
# "Pvalue" - non-adjusted p-value; "OddsRatio" - enrichment odds ratio; "ExpCount" - number of genes 
# expected to be selected in a category; "Count" - number of genes observed in the current list; 
# "Size" - total number of genes in a category; "Term" - category description; "p.adj" - false  
# discovery rate; "SYMBOL", "ENTREZ" - genes observed in the current list as annotated with a category 

# To perform Reactome and Disease Ontology enrichment analyses, use the corresponding packages
library("ReactomePA")
library("DOSE")
res <- enrichPathway(gene = entrez.genes, universe = all.entrez, organism = "human", pAdjustMethod = "BH", pvalueCutoff = 0.1, qvalueCutoff = 0.2, readable = TRUE)
write.xlsx(summary(res), fileName, sheetName = "Reactome",row.names=FALSE,  append = TRUE)
res <- enrichDO(gene = entrez.genes, universe = all.entrez, ont = "DO", pAdjustMethod = "none", pvalueCutoff = 0.1, qvalueCutoff = 0.2, readable = TRUE)
write.xlsx(summary(res), fileName, sheetName = "DiseaseOntology",row.names=FALSE,  append = TRUE)
}

}

