% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_gene_extract.R
\name{gr_gene_extract}
\alias{gr_gene_extract}
\title{Extract gene regions for the list of genes}
\usage{
gr_gene_extract(selected, id = "entrezid")
}
\arguments{
\item{selected}{a character vector of gene Entrez IDs, or names. Required}

\item{id}{what type of ID is provided. "symbol" (e.g., "BRCA1") or "entrezid" (e.g., "672", default and recommended).}
}
\value{
a list with two components. 'promoters' contains genomic coordinates
of the selected genes; 'notfound' contains gene names not found in annotables (for diagnostics)
}
\description{
A function to extract genomic coordinates of genes in BED format.
}
\details{
Only canonical chromosome names are considered. No patches, random or other contigs.
}
\examples{
\dontrun{
selected <- c("TMEM59L", "SNTG1", "RPL41", "ADAMTS19")
pr <- gr_promoter_extract(selected)$promoters
write.table(pr, "pr.bed", sep="\\t", quote=F, row.names=F, col.names=F)
# Extract promoters of all genes (that have gene names). Note only canonical chromosomes are retained.
promoters.all <- gr_promoter_extract(selected = unique(annotables::grch37$symbol))
}
}

