% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gr_maxmin.R
\name{gr_maxmin}
\alias{gr_maxmin}
\title{Extract pairs of objects from a correlation matrix, sorted by max/min
correlation coefficient}
\usage{
gr_maxmin(mtx, cortype = "spearman", fileName = NULL)
}
\arguments{
\item{mtx}{a n x m matrix used to create an m x m square matrix of correlation
coefficients (column-wise correlations). Required}

\item{cortype}{method to calculate correlation coefficients. Default - "spearman".
Other oprions - "pearson".}

\item{fileName}{name of a tab-separated text file to save the results.
Default - none, do not save the results}
}
\value{
a 4-column matrix of object.1, object.2, correlation.coefficient,
p.value, sorted by max-to-min correlation coefficient. object.1, 2 etc. are
the names of the columns of the original matrix.
}
\description{
A function to extract pairs of objects from a square correlation matrix,
sorted by max/min correlation coefficient. The self-self correlations
are excluded.
}
\examples{
\dontrun{
mtx.maxmin <- gr_maxmin(mtx, cortype="pearson", fileName = "maxmin.txt")
}
}

