% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_promoter_extract.R
\name{gr_promoter_extract}
\alias{gr_promoter_extract}
\title{Extract promoter regions for the list of genes}
\usage{
gr_promoter_extract(selected, upstream = 2000, downstream = 500)
}
\arguments{
\item{selected}{a character vector of gene names (EntrezIDs support in progress). Required}

\item{upstream}{how many base pairs upstream of TSS define promoter. Default - 2,000bp.}

\item{downstream}{how many base pairs downstream of TSS define promoter. Default - 500bp.}
}
\value{
a list with two components. 'promoters' contains genomic coordinates
of the selected genes; 'notfound' contains gene names not found in annotables (for diagnostics)
}
\description{
A function to extract genomic coordinates of gene promoters in BED format.
}
\details{
Only canonical chromosome names are considered. No patches, random or other contigs.
}
\examples{
\dontrun{
selected <- c("TMEM59L", "SNTG1", "RPL41", "ADAMTS19")
pr <- gr_promoter_extract(selected)$promoters
write.table(pr, "pr.bed", sep="\\t", quote=F, row.names=F, col.names=F)
# Extract promoters of all genes (that have gene names). Note only canonical chromosomes are retained.
promoters.all <- gr_promoter_extract(selected = unique(annotables::grch37$symbol))
}
}

