% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_untransform.R
\name{gr_untransform}
\alias{gr_untransform}
\title{Conversion of a matrix of -log10-transformed p-values into a linear scale}
\usage{
gr_untransform(mtx, p2z = FALSE)
}
\arguments{
\item{mtx}{a matrix of -log10-transformed p-values, 
with "-" sign indicating depleted enrichments. Required.}

\item{p2z}{logical. Indicates whether instead of -log10-transformation use
Z-scores. Default - FALSE.}
}
\value{
a matrix of untransformed p-values, with "-" sign preserved
}
\description{
A function to convert a matrix of -log10-transformed p-values (with "-" 
indicating depletion) into a linear scale
}
\examples{
\dontrun{
mtx <- gr_untransform(mtx.transformed)
}
}

