% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_corr.R
\name{TCGA_corr}
\alias{TCGA_corr}
\title{Retrieves correlation statistics between a gene of interest and all other genes.}
\usage{
TCGA_corr(canter = "BRCA", gene = "BRCA1")
}
\arguments{
\item{gene}{gene of interest. Required. Example: "BRCA1"}

\item{cancer}{- cancer type abbreviation. Required. Example: "BRCA". Look up
abbreviations [here](http://www.liuzlab.org/TCGA2STAT/CancerDataChecklist.pdf)
and [here](https://tcga-data.nci.nih.gov/datareports/codeTablesReport.htm)}
}
\value{
a sorted by correlation coefficient data frame. Columns are 
"hgnc_symbol", "cor", "pval", "description"
}
\description{
A function to get Pearson correlation coefficients between RSEM gene expression of
a gene of interest and all other genes in a specific cancer type.
}
\examples{
\dontrun{
BRCA1_corr <- TCGA_corr(canter = "BRCA", gene = "BRCA1")
}
}

